+++ web/src/components/cards/Missions.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable max-lines -- TODO: split this file (tracked by #15790) */
 import { useState, useEffect, useRef, useMemo } from 'react'
 import {
   Rocket,
